Move TraversersTest to correct package
